requirements.txt
discord.py-self
Jinja2
aiofiles
markdown